refactor(commit): コマンド処理を関数化し再利用性を向上
- konokaEdit関数を新規追加しコマンド処理本体を分離
- bin/konoka-editor.tsのコマンド実装を簡素化しエントリーポイントをmain関数化
- 型定義やコメントを整理し可読性を向上
